+++ examples/image_streamer/deployment_plan.rb
@@ -28,7 +28,7 @@ image_streamer_deployment_plan 'DeploymentPlan1' do
     description: 'AnyDescription',
     hpProvided: false
   )
-  build_plan 'ChefBP01'
+  os_build_plan 'ChefBP01'
   golden_image 'ChefGI01'
 end
 
@@ -49,7 +49,7 @@ image_streamer_deployment_plan 'DeploymentPlan2' do
     description: 'example of create_if_missing action',
     hpProvided: false
   )
-  build_plan 'ChefBP01'
+  os_build_plan 'ChefBP01'
   golden_image 'ChefGI01'
   action :create_if_missing
 end