+++ libraries/resource_providers/image_streamer/api300/deployment_plan.rb
@@ -17,14 +17,14 @@ module OneviewCookbook
       # DeploymentPlan Provider resource methods
       class DeploymentPlanProvider < ResourceProvider
         def create_or_update
-          @item['goldenImageURI'] ||= load_resource(:GoldenImage, @context.golden_image)
-          @item['oeBuildPlanURI'] ||= load_resource(:BuildPlan, @context.build_plan)
+          @item['goldenImageURI'] ||= load_resource(:GoldenImage, @context.golden_image, :uri)
+          @item['oeBuildPlanURI'] ||= load_resource(:BuildPlan, @context.os_build_plan, :uri)
           super
         end
 
         def create_if_missing
-          @item['goldenImageURI'] ||= load_resource(:GoldenImage, @context.golden_image)
-          @item['oeBuildPlanURI'] ||= load_resource(:BuildPlan, @context.build_plan)
+          @item['goldenImageURI'] ||= load_resource(:GoldenImage, @context.golden_image, :uri)
+          @item['oeBuildPlanURI'] ||= load_resource(:BuildPlan, @context.os_build_plan, :uri)
           super
         end
       end